modified ddl types for creating source tables

+++ dags/sql_scripts/src_accounts.sql
@@ -1,7 +1,7 @@
 -- "account_id" string;"district_id"int;"frequency" low cardinality string;"date" date yymmdd
 CREATE TABLE IF NOT EXISTS src_accounts (
     account_id String,
-    district_id Int32,
+    district_id UInt64,
     frequency LowCardinality(String),
     date Date
 ) ENGINE = MergeTree()
