+++ dags/sql_scripts/src_cards.sql
@@ -1,7 +1,7 @@
 -- "card_id" int;"disp_id" int;"type" low cardinality string;"issued" date time yymmdd hh:mm:ss
 CREATE TABLE IF NOT EXISTS src_cards (
-    card_id Int32,
-    disp_id Int32,
+    card_id UInt64,
+    disp_id UInt64,
     type LowCardinality(String),
     issued DateTime
 ) ENGINE = MergeTree()
