+++ dags/sql_scripts/src_disposition.sql
@@ -1,7 +1,7 @@
 -- "disp_id" int;"client_id" int;"account_id" string;"type" low cardinality string
 CREATE TABLE IF NOT EXISTS src_disposition (
-    disp_id Int32,
-    client_id Int32,
+    disp_id UInt64,
+    client_id UInt64,
     account_id String,
     type LowCardinality(String)
 ) ENGINE = MergeTree()
