+++ dags/sql_scripts/src_permanent_orders.sql
@@ -1,6 +1,6 @@
 -- "order_id" int;"account_id" string;"bank_to" string;"account_to" string;"amount" float;"k_symbol" low cardinality string
 CREATE TABLE IF NOT EXISTS src_permanent_orders (
-    order_id Int32,
+    order_id UInt64,
     account_id String,
     bank_to String,
     account_to String,
